pyproject.toml
[project]
name = "aic-community-collector"
version = "0.1.0"
description = "AIC 챌린지 커뮤니티 데이터 수집 도구"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.30",
    "pyyaml",
    "numpy",
    "scipy",
    "prefect>=3.0",
    "fakeredis<2.27",
]

[project.scripts]
aic-collector = "aic_collector.webapp:main"
aic-prefect-run = "aic_collector.prefect.serve:main"
aic-collector-worker = "aic_collector.job_queue.consumer_cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/aic_collector"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
